* commit '2b5a41ae082b07f26ef871fd71ee742f519576df':
  PROJECTFORGE-2915: Bug fixed

+++ projectforge-wicket/src/main/java/org/projectforge/web/vacation/VacationEditPage.java
@@ -134,22 +134,27 @@ public class VacationEditPage extends AbstractEditPage<VacationDO, VacationEditF
       } else if (VacationStatus.IN_PROGRESS == form.getData().getStatus()) {
         vacationService.sendMailToVacationInvolved(form.getData(), false, false);
       }
+      if (form.getStatusBeforeModification() != null) {
+        if (form.getStatusBeforeModification().equals(VacationStatus.APPROVED) && (VacationStatus.REJECTED.equals(form.getData().getStatus())
+            || VacationStatus.IN_PROGRESS.equals(form.getData().getStatus()))) {
+          vacationService.deleteUsedVacationDaysFromLastYear(form.getData());
+        }
+        if (form.getStatusBeforeModification() == VacationStatus.IN_PROGRESS) {
+          switch (form.getData().getStatus()) {
+            case APPROVED:
+              // IN_PROGRESS -> APPROVED
+              vacationService.updateUsedVacationDaysFromLastYear(form.getData());
+              vacationService.sendMailToEmployeeAndHR(form.getData(), true);
+              break;
 
-      if (form.getStatusBeforeModification() == VacationStatus.IN_PROGRESS) {
-        switch (form.getData().getStatus()) {
-          case APPROVED:
-            // IN_PROGRESS -> APPROVED
-            vacationService.updateUsedVacationDaysFromLastYear(form.getData());
-            vacationService.sendMailToEmployeeAndHR(form.getData(), true);
-            break;
+            case REJECTED:
+              // IN_PROGRESS -> REJECTED
+              vacationService.sendMailToEmployeeAndHR(form.getData(), false);
+              break;
 
-          case REJECTED:
-            // IN_PROGRESS -> REJECTED
-            vacationService.sendMailToEmployeeAndHR(form.getData(), false);
-            break;
-
-          default:
-            // nothing to do
+            default:
+              // nothing to do
+          }
         }
       }
     } catch (final Exception e) {
@@ -163,8 +168,10 @@ public class VacationEditPage extends AbstractEditPage<VacationDO, VacationEditF
   public WebPage afterDelete()
   {
     try {
-      vacationService.deleteUsedVacationDaysFromLastYear(form.getData());
-      vacationService.sendMailToVacationInvolved(form.getData(), false, true);
+      if (VacationStatus.APPROVED.equals(form.getData().getStatus())) {
+        vacationService.deleteUsedVacationDaysFromLastYear(form.getData());
+        vacationService.sendMailToVacationInvolved(form.getData(), false, true);
+      }
     } catch (final Exception e) {
       log.error("There is a exception in afterDelete: " + e.getMessage(), e);
       error(I18nHelper.getLocalizedMessage("vacation.error.sendmail"));
@@ -175,8 +182,12 @@ public class VacationEditPage extends AbstractEditPage<VacationDO, VacationEditF
   public WebPage afterUndelete()
   {
     try {
-      vacationService.updateUsedVacationDaysFromLastYear(form.getData());
-      vacationService.sendMailToVacationInvolved(form.getData(), false, false);
+      if (VacationStatus.APPROVED.equals(form.getData().getStatus())) {
+        vacationService.updateUsedVacationDaysFromLastYear(form.getData());
+        vacationService.sendMailToEmployeeAndHR(form.getData(), true);
+      } else {
+        vacationService.sendMailToVacationInvolved(form.getData(), false, false);
+      }
     } catch (final Exception e) {
       log.error("There is a exception in afterUndelete: " + e.getMessage(), e);
       error(I18nHelper.getLocalizedMessage("vacation.error.sendmail"));
